Finished claims part

+++ section 6 sql-alchemy/Code/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 from models.item import ItemModel
 
@@ -37,12 +37,17 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted'}
+        return {'message': 'Item not found'}, 404
 
     # We only allow updating of price. Not the name of the item
     def put(self, name):
